Specialization route and produces response type handled properly with exception handling

+++ ClinicCare.Business/Services/PatientService.cs
@@ -1,17 +1,21 @@
 ﻿using ClinicCare.Business.Exceptions;
+using ClinicCare.Business.Interfaces;
 using ClinicCare.Business.Services.Interfaces;
 using ClinicCare.DataAccess.Models;
 using ClinicCare.DataAccess.Repositories.Interfaces;
 using ClinicCare.Shared.DTOs.Patient;
+using ClinicCare.Shared.Enums;
 
 namespace ClinicCare.Business.Services
 {
     public class PatientService : IPatientService
     {
         private readonly IGenericRepository<Patient> _repo;
+        private readonly ICurrentUser _currentUser;
 
-        public PatientService(IGenericRepository<Patient> repo)
+        public PatientService(IGenericRepository<Patient> repo, ICurrentUser currentUser)
         {
+            _currentUser = currentUser;
             _repo = repo;
         }
 
@@ -32,9 +36,11 @@ namespace ClinicCare.Business.Services
         public async Task<PatientResponseDto?> GetByIdAsync(Guid id)
         {
             var patient = await _repo.GetByIdAsync(id);
-
             if (patient is null)
                 throw new NotFoundException($"Patient with id {id} not found.");
+
+            if (_currentUser.Role == UserRole.Patient && _currentUser.UserId != id)
+                throw new ForbiddenException("You are not authorized");
 
             return new PatientResponseDto
             {
@@ -49,9 +55,11 @@ namespace ClinicCare.Business.Services
         public async Task UpdateAsync(Guid id, PatientUpdateDto dto)
         {
             var patient = await _repo.GetByIdAsync(id);
-
             if (patient is null)
                 throw new NotFoundException($"Patient with id {id} not found.");
+
+            if (_currentUser.Role == UserRole.Patient && _currentUser.UserId != id)
+                throw new ForbiddenException("You are not authorized");
 
             patient.FirstName = dto.FirstName;
             patient.LastName = dto.LastName;
@@ -65,9 +73,11 @@ namespace ClinicCare.Business.Services
         public async Task DeleteAsync(Guid id)
         {
             var patient = await _repo.GetByIdAsync(id);
-
             if (patient is null)
                 throw new NotFoundException($"Patient with id {id} not found.");
+
+            if (_currentUser.Role == UserRole.Patient && _currentUser.UserId != id)
+                throw new ForbiddenException("You are not authorized");
 
             await _repo.Delete(id);
             await _repo.SaveChangesAsync();
